% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{draw a plot with a name as input}
\usage{
draw_a_name(a_name)
}
\arguments{
\item{a_name}{}
}
\value{
ggplot graph
}
\description{
draw a plot with a name as input
}
