% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save data frame into a csv file}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
invisible("/path/of/saved/csv)
}
\description{
Save data frame into a csv file
}
